+++ modules/core/src/main/scala/com/odenzo/xrpl/signing/core/models/XrpKeyPair.scala
@@ -1,0 +1,3 @@
+package com.odenzo.xrpl.signing.core.models
+
+case class XrpKeyPair[PRIVATE, PUBLIC](privateKey: PRIVATE, publicKey: PUBLIC)