+++ modules/core/src/main/scala/com/odenzo/xrpl/signing/models/XrpKeyPair.scala
@@ -1,3 +1,0 @@
-package com.odenzo.xrpl.signing.models
-
-case class XrpKeyPair[PRIVATE, PUBLIC](privateKey: PRIVATE, publicKey: PUBLIC)