+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
-## V0.0.1
+## v0.0.1
 Current working version, no releases. Being revamped and refactored. Might put this back 
 into a more heavy-weight program that uses the main models and codecs to product hash and tx_json
+
+## v0.0.2
+
+- Stablize compile
+- Test RFC1751 "phrase" to XrpSeed
+-